Still improving drive-trhu display, bug fixed in calculation of the width
- now correct width is measured (calculated)
- ini file updated to a new distance (easier to test with)
- cursor in laserView (GUI) changed to default
- comments added
- fixme tags added (still found some fix 180 degree values in some loops!)

+++ direcs/src/obstacleCheckThread.h
@@ -116,10 +116,10 @@ class ObstacleCheckThread : public QThread
 		
 		int centerOfFreeWay;
 		
-		float width;
-		float b;
-		float c;
-		float alpha;
+		double width;
+		double b;
+		double c;
+		double alpha;
 		
 		/*! defines the minimum slot in degrees, where the robot has to fit through */
 		int robotSlot;
